+++ src/player.h
@@ -1179,8 +1179,8 @@ class Player final : public Creature, public Cylinder
 		void removeThing(Thing* thing, uint32_t count) final;
 
 		int32_t getThingIndex(const Thing* thing) const final;
-		int32_t getFirstIndex() const final;
-		int32_t getLastIndex() const final;
+		size_t getFirstIndex() const final;
+		size_t getLastIndex() const final;
 		uint32_t getItemTypeCount(uint16_t itemId, int32_t subType = -1) const final;
 		std::map<uint32_t, uint32_t>& getAllItemTypeCount(std::map<uint32_t, uint32_t> &countMap) const final;
 		Thing*getThing(size_t index) const final;
